+++ HISTORY.md
@@ -5,10 +5,17 @@ This file will contain current and previous changes to Crackle. Make sure to upd
 
 # In development
 
+## 2025-11-10
+ * `categoryCreated` event to match `categoryCreating`
+ 
 ## 2025-11-09
  * better API doc
- * events - only 'projectCreating' for now
- 
+ * events - only `projectCreating` for now
+ * new events:
+    * `projectCreated`
+    * `categoryCreating`
+ * change `attachEventHandlers` to have the ide be passed in
+
 ## 2025-11-08
  * create HISTORY.md file like Snap!
  * helpers file not really needed - move to index.js